LPS-79037 change packages to addapt to the new naming convention
