+++ modules/apps/headless-apio/category/category-apio-api/bnd.bnd
@@ -1,4 +1,4 @@
 Bundle-Name: Liferay Category Apio API
 Bundle-SymbolicName: com.liferay.category.apio.api
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.category.apio.identifier
+Export-Package: com.liferay.category.apio.identifier.architect
