+++ modules/apps/headless-apio/category/category-apio-api/src/main/java/com/liferay/category/apio/identifier/architect/CategoryIdentifier.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.category.apio.identifier;
+package com.liferay.category.apio.identifier.architect;
 
 import com.liferay.apio.architect.identifier.Identifier;
