+++ modules/apps/headless-apio/category/category-apio-api/src/main/resources/com/liferay/category/apio/identifier/architect/packageinfo
